Refactor SWG events listing component to more general SWG Events component
This will eventually have different views for different tasks,
e.g. leaders submitting walks, assembling a walks programme, etc.

There is a database change to go along with this commit:
Any previous occurrences of swg_eventlisting must be changed to
swg_events (if referring to the module) or
eventlisting (if referring to the view)
Alternatively, delete and recreate everything that uses the component.