Change JS map route handling to handle routes and tracks separately
In event listings, a point-to-point route is a dotted line, a planned route is a dashed line,
and a recorded track is a solid line.

Route::loadForWalkable takes extra parameter for the route type - planned or logged

SWGMap provides a loadedHandler, which triggers when all events have finished loading.
Pass a function to addLoadedHandler to use it. By default a handler triggers once -
pass boolean true to repeat the handler.