Major changes to how events are handled in the API and JS
The events listing page is now more responsible for its API parameters.
This allows it to modify the start and end date to carry on from the
events already loaded. All events on the events listing page and in the
next event popup now have at least a partial Javascript object
representation, which makes comparison much easier. Events on the page
during the initial load use machine-readable markup to get important
details: type, id, start & end dates, title. Events loaded via JSON just
store the data received in full. Start & end dates are converted to JS
date objects.

Leader contact details are removed from walks in the past before being
sent to the API. The method is fairly crude, but works for now.

Basic hCalendar support has been added to the event listing page. This
is currently mostly used so we have machine readable dates for JS, but
it should work if anyone wants to try it. End dates are not currently
added to events generated by JS, because end dates come in various
different formats that should be unified in PHP.

Users can now select if they went to an event. This will tie in with the
future walk diary functionality, but currently is displayed to all users
even if they're not logged in - although the data cannot be stored. This
will require a new permission to be defined. If the current user said
they went to a walk, there are extra links inviting them to submit a GPS
track and photos.

Some parts of the event listing HTML construction has been moved out of
the common events.js file and into nextevents.js. This is to stop it
being used for past events, where we display less information, but this
needs revising because theoretically we could use a bottomless page to
load future events.