Fix display of attendee count and 'you did this' in event listing
Looks like I'd changed the way attendees were stored on Events, but
hadn't changed the factories. Both sides have been updated so the
factories can load attendance status for a user (normally the current
user) and send it to the Event, which caches it. The event can then
use this stored data, or perform its own queries if it needs something
else.