Fix bottomless page on events listing:
There was a walk in August 2013 with no details set (not even a name).
Events with no name are now not sent out by the event listing (PHP).
When JS stores & displays an event, it now assumes attendance is an array.
Cancelled events have no 'You did this' box, and have the 'cancelled' mark.