Big overhaul of mobile site, especially homepage
We now have a full style for the mobile homepage, showing the slideshow
and new members' info, but without the upcoming events. The main heading
on every page has been put above the slideshow (larger on the homepage),
and the second heading on the homepage is below the slideshow. The
second heading on other pages is still on the slideshow, but larger and
nearer the top.

The menu bar is now slightly smaller - 48px high, seems a good size on
my phone. Slideshow images are resized to fit the screen, and there's a
bit of padding at the top of the footer.