Basic page for showing details about a specific event
Will be used to show more info than can be shown on the event listing,
e.g. who went, photos, comments, etc.

Currently shows the number of people who logged the event in their diary
(but this doesn't include the organisers).

Link in header is removed on the mobile site because it interferes with
tapping on the header to expand the event