Update user stats on screen when changing attendance
The AJAX call to record attendance for an event now returns the user
stats in JSON format. This is passed by event and picked up by both the
user stats module and the profile module, which now update immediately.
The stats helper includes converting the distance to miles, which is
defined in one place only to make it easier to change in future.

Fixed capitalisation of distance & displayDistance methods in unit
converter